# doc/api.yml
openapi: "3.0.3"
info:
  title: Bookstore API
  version: "1.0.0"
  description: >
    REST API for the onlydevs bookstore digitalization platform.
    Covers catalog management, inventory tracking, and point-of-sale.
servers:
  - url: /api/v1
    description: Default server
...
components:
  schemas:
...
    AuthorRole:
      type: string
      enum: [PRIMARY, CO_AUTHOR, EDITOR, TRANSLATOR, ILLUSTRATOR, FOREWORD]
...
    CreateAuthorRequest:
      type: object
      required: [firstName, lastName]
      properties:
        firstName: { type: string, maxLength: 100 }
        lastName: { type: string, maxLength: 100 }
    UpdateAuthorRequest:
      type: object
      properties:
        firstName: { type: string, maxLength: 100 }
        lastName: { type: string, maxLength: 100 }
    AddBookAuthorRequest:
      type: object
      required: [role]
      properties:
        role: { $ref: '#/components/schemas/AuthorRole' }
        contributionOrder: { type: integer, minimum: 1 }
...
    AuthorResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        firstName: { type: string }
        lastName: { type: string }
        fullName: { type: string }
        createdAt: { type: string, format: date-time }
    BookAuthorResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        bookId: { type: string, format: uuid }
        bookTitle: { type: string }
        authorId: { type: string, format: uuid }
        authorFullName: { type: string }
        role: { $ref: '#/components/schemas/AuthorRole' }
        contributionOrder: { type: integer }
...
paths:
...
  /books/{bookId}/authors/{authorId}:
    post:
      tags: [BookAuthors]
      summary: Add author to book with role
      parameters:
        - in: path
          name: bookId
          required: true
          schema: { type: string, format: uuid }
        - in: path
          name: authorId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddBookAuthorRequest' }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BookAuthorResponse' }
        '409': { description: Author already linked to this book }
    delete:
      tags: [BookAuthors]
      summary: Remove author from book
      parameters:
        - in: path
          name: bookId
          required: true
          schema: { type: string, format: uuid }
        - in: path
          name: authorId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Removed }
        '400': { description: Cannot remove last author }
...
  /authors:
    get:
      tags: [Authors]
      summary: Search authors by name
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuthorResponse'
    post:
      tags: [Authors]
      summary: Create author
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAuthorRequest' }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthorResponse' }
  /authors/{id}:
    get:
      tags: [Authors]
      summary: Get author detail
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthorResponse' }
    put:
      tags: [Authors]
      summary: Update author
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateAuthorRequest' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthorResponse' }
    delete:
      tags: [Authors]
      summary: Delete author
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses: { '204': { description: Deleted } }

  # Book Editions
...